+++ app/build.gradle
@@ -45,7 +45,6 @@ dependencies {
     annotationProcessor 'com.jakewharton:butterknife-compiler:8.8.1'
     implementation 'com.android.support:design:27.0.2'
 
-
     // LeanCloud 基础包
     compile('cn.leancloud.android:avoscloud-sdk:v4.6.4')
