--- app/src/main/java/com/example/nesthabit/base/ItemOnClickListener.java
@@ -1,7 +1,0 @@
-package com.example.nesthabit.base;
-
-import android.view.View;
-
-public interface ItemOnClickListener {
-    void onItemClick(View view, int position);
-}
