--- app/src/main/java/com/example/nesthabit/fragment/ClockSetView.java
@@ -1,6 +1,0 @@
-package com.example.nesthabit.clock_set;
-
-import com.example.nesthabit.base.BaseView;
-
-public interface ClockView extends BaseView {
-}
