--- app/src/main/java/com/example/nesthabit/fragment/ClockView.java
@@ -1,7 +1,0 @@
-package com.example.nesthabit.clock;
-
-import com.example.nesthabit.base.BaseView;
-
-public interface ClockView extends BaseView {
-
-}
