+++ app/src/main/java/com/example/nesthabit/home/HomePresenter.java
@@ -1,6 +1,5 @@
 package com.example.nesthabit.home;
 
-import com.example.nesthabit.activity.HomeView;
 import com.example.nesthabit.base.BasePresent;
 
 public class HomePresenter extends BasePresent<HomeView> {
