--- app/src/main/java/com/example/nesthabit/model/bean/ClockInfo.java
@@ -1,7 +1,0 @@
-package com.example.nesthabit.model.bean;
-
-public class ClockInfo {
-    public boolean isOpen;
-    public String slogan;
-    public boolean[] days;
-}
