--- app/src/main/java/com/example/nesthabit/model/temp.java
@@ -1,4 +1,0 @@
-package com.example.nesthabit.model;
-
-public class temp {
-}
