--- app/src/main/java/com/example/nesthabit/presenter/ClockFraPresenter.java
@@ -1,7 +1,0 @@
-package com.example.nesthabit.clock;
-
-import com.example.nesthabit.base.BasePresent;
-import com.example.nesthabit.fragment.ClockView;
-
-public class ClockFraPresenter extends BasePresent<ClockView> {
-}
