--- app/src/main/java/com/example/nesthabit/presenter/ClockSetPresenter.java
@@ -1,8 +1,0 @@
-package com.example.nesthabit.clock_set;
-
-import com.example.nesthabit.base.BasePresent;
-import com.example.nesthabit.fragment.ClockSetView;
-
-public class ClockSetPresenter extends BasePresent<ClockSetView> {
-
-}
