--- app/src/main/res/layout/activity_clock_set.xml
@@ -1,137 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:card_view="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="#FF8A85"
-        android:orientation="vertical">
-
-        <android.support.v7.widget.Toolbar
-            android:id="@+id/clock_set_toolbar"
-            android:layout_width="match_parent"
-            android:layout_height="56dp">
-
-        </android.support.v7.widget.Toolbar>
-
-        <android.support.v7.widget.CardView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="10dp"
-            card_view:cardCornerRadius="12dp">
-
-            <TimePicker
-                android:id="@+id/clock_set_time_picker"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:timePickerMode="spinner"
-                android:textColor="#5c6b73" />
-
-        </android.support.v7.widget.CardView>
-
-        <android.support.v7.widget.CardView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            card_view:cardBackgroundColor="#ffffff"
-            card_view:cardCornerRadius="12dp"
-            android:layout_margin="10dp">
-
-            <RelativeLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_margin="10dp">
-
-                <TextView
-                    android:id="@+id/clock_set_1"
-                    android:layout_width="80dp"
-                    android:layout_height="wrap_content"
-                    android:text="标题"
-                    android:textSize="20sp"
-                    android:textColor="#5c6b73"
-                    android:layout_margin="5dp" />
-
-                <EditText
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_toRightOf="@id/clock_set_1"
-                    android:background="@null"
-                    android:text="我要早起"
-                    android:textAlignment="viewEnd"
-                    android:textColor="#5c6b73"
-                    android:textSize="20sp"
-                    android:layout_margin="5dp" />
-
-                <View
-                    android:id="@+id/clock_set_divider1"
-                    android:layout_below="@id/clock_set_1"
-                    android:layout_width="fill_parent"
-                    android:layout_height="1px"
-                    android:background="#5c6b73" />
-
-                <TextView
-                    android:id="@+id/clock_set_2"
-                    android:layout_width="80dp"
-                    android:layout_height="wrap_content"
-                    android:layout_below="@id/clock_set_divider1"
-                    android:text="重复"
-                    android:textSize="20sp"
-                    android:layout_margin="5dp"
-                    android:textColor="#5c6b73" />
-
-                <include
-                    layout="@layout/days_button"
-                    android:layout_margin="5dp"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_below="@id/clock_set_divider1"
-                    android:layout_toRightOf="@id/clock_set_2" />
-
-                <View
-                    android:id="@+id/clock_set_divider2"
-                    android:layout_below="@id/clock_set_2"
-                    android:layout_width="fill_parent"
-                    android:layout_height="1px"
-                    android:background="#5c6b73" />
-
-                <TextView
-                    android:id="@+id/clock_set_3"
-                    android:layout_width="80dp"
-                    android:layout_height="wrap_content"
-                    android:text="对应鸟窝"
-                    android:textColor="#5c6b73"
-                    android:textSize="20sp"
-                    android:layout_margin="5dp"
-                    android:layout_below="@id/clock_set_divider2" />
-
-                <LinearLayout
-                    android:layout_width="wrap_content"
-                    android:orientation="horizontal"
-                    android:layout_height="wrap_content"
-                    android:layout_below="@id/clock_set_divider2"
-                    android:layout_toRightOf="@id/clock_set_3">
-
-                    <TextView
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:textColor="#5c6b73"
-                        android:textSize="20sp"
-                        android:layout_margin="5dp"
-                        android:text="早起的鸟儿有虫吃" />
-
-                    <Button
-                        android:layout_width="20dp"
-                        android:layout_height="20dp"
-                        android:background="@mipmap/right" />
-
-                </LinearLayout>
-
-
-            </RelativeLayout>
-
-
-        </android.support.v7.widget.CardView>
-    </LinearLayout>
-</ScrollView>
