--- app/src/main/res/layout/activity_home.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <android.support.v7.widget.Toolbar
-        android:id="@+id/home_toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="56dp"
-        android:layout_alignParentTop="true">
-
-    </android.support.v7.widget.Toolbar>
-
-
-    <android.support.v4.view.ViewPager
-        android:id="@+id/home_view_pager"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_above="@id/bottom"
-        android:layout_below="@id/home_toolbar" />
-
-    <include
-        layout="@layout/home_bottom"
-        android:layout_width="match_parent"
-        android:layout_height="56dp"
-        android:layout_alignParentBottom="true"
-        android:id="@+id/bottom" />
-
-
-</RelativeLayout>
