--- app/src/main/res/layout/clock_item.xml
@@ -1,88 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.v7.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_margin="5dp"
-    android:padding="5dp">
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <ImageView
-            android:id="@+id/clock_item_img"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentLeft="true"
-            android:src="@mipmap/night" />
-
-        <RelativeLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_toEndOf="@id/clock_item_img"
-            android:layout_toRightOf="@id/clock_item_img"
-            android:padding="5dp"
-            android:id="@+id/clock_item_checked">
-
-            <TextView
-                android:layout_alignParentTop="true"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:id="@+id/clock_item_slogan"
-                android:layout_margin="5dp"
-                android:textSize="25sp"
-                android:textColor="#5c6b73"
-                android:text="我要早起" />
-
-            <ImageView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:id="@+id/clock_item_logo"
-                android:layout_toRightOf="@id/clock_item_slogan"
-                android:background="#ffffff"
-                android:src="@mipmap/friend_remind_none"
-                android:layout_margin="5dp" />
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:id="@+id/clock_item_time"
-                android:layout_below="@id/clock_item_slogan"
-                android:layout_margin="5dp"
-                android:textSize="45sp"
-                android:textColor="#5c6b73"
-                android:text="07:59"
-                android:textStyle="bold" />
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_margin="5dp"
-                android:textSize="25sp"
-                android:textColor="#5c6b73"
-                android:id="@+id/clock_item_day"
-                android:layout_below="@id/clock_item_time"
-                android:text="每天" />
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_margin="5dp"
-                android:textSize="25sp"
-                android:textColor="#78909c"
-                android:id="@+id/clock_item_lefttime"
-                android:layout_below="@id/clock_item_day"
-                android:text="还有23小时60分钟" />
-
-
-        </RelativeLayout>
-
-        <Switch
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_alignParentEnd="true"
-            android:id="@+id/clock_item_switch" />
-
-    </RelativeLayout>
-
-</android.support.v7.widget.CardView>
