--- app/src/main/res/layout/days_button.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_gravity="center_vertical"
-    android:gravity="center_vertical"
-    android:orientation="horizontal"
-    android:layout_margin="5dp">
-
-    <Button
-        android:id="@+id/clock_days_0"
-        android:layout_width="30dp"
-        android:layout_height="30dp"
-        android:background="@drawable/clock_button" />
-
-    <Button
-        android:id="@+id/clock_days_1"
-        android:layout_width="30dp"
-        android:layout_height="30dp"
-        android:background="@drawable/clock_button" />
-
-    <Button
-        android:id="@+id/clock_days_2"
-        android:layout_width="30dp"
-        android:layout_height="30dp"
-        android:background="@drawable/clock_button" />
-
-    <Button
-        android:id="@+id/clock_days_3"
-        android:layout_width="30dp"
-        android:layout_height="30dp"
-        android:background="@drawable/clock_button" />
-
-    <Button
-        android:id="@+id/clock_days_4"
-        android:layout_width="30dp"
-        android:layout_height="30dp"
-        android:background="@drawable/clock_button" />
-
-    <Button
-        android:id="@+id/clock_days_5"
-        android:layout_width="30dp"
-        android:layout_height="30dp"
-        android:background="@drawable/clock_button" />
-
-    <Button
-        android:id="@+id/clock_days_6"
-        android:layout_width="30dp"
-        android:layout_height="30dp"
-        android:background="@drawable/clock_button" />
-
-    <Button
-        android:id="@+id/clock_days_7"
-        android:layout_width="30dp"
-        android:layout_height="30dp"
-        android:background="@drawable/clock_button" />
-
-</LinearLayout>
